+++ dart/test/protocol/contexts_test.dart
@@ -22,6 +22,7 @@ void main() {
     final app = SentryApp(name: 'name1');
     final browser = SentryBrowser(name: 'name1');
     final gpu = SentryGpu(id: 2);
+    final culture = SentryCulture(locale: 'foo-bar');
 
     final copy = data.copyWith(
       device: device,
@@ -30,6 +31,7 @@ void main() {
       app: app,
       browser: browser,
       gpu: gpu,
+      culture: culture,
     );
 
     expect(device.toJson(), copy.device!.toJson());
@@ -40,6 +42,7 @@ void main() {
     );
     expect(app.toJson(), copy.app!.toJson());
     expect(browser.toJson(), copy.browser!.toJson());
+    expect(culture.toJson(), copy.culture!.toJson());
     expect(gpu.toJson(), copy.gpu!.toJson());
     expect('value', copy['extra']);
   });
@@ -52,4 +55,5 @@ Contexts _generate() => Contexts(
       app: SentryApp(name: 'name'),
       browser: SentryBrowser(name: 'name'),
       gpu: SentryGpu(id: 1),
+      culture: SentryCulture(locale: 'foo-bar'),
     );