+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 * Feat: Enrich events with more context (#452)
+* Feat: Add Culture Context (#491)
+* Remove `Sentry.currentHub` (#490)
 
 # 5.1.0